Stash some updates before change of workstream
