+++ src/worker/code.ts
@@ -1,142 +1,51 @@
-import { chamfer, fillet, move, rotate, scale } from "./actions";
-import { ReplicadObject } from "./geometryProvider";
-import { assembly, cutAssembly, intersect } from "./interaction";
+/**
+ * Code atom executor dispatcher.
+ *
+ * This file is the single entry point from the worker into user-authored
+ * Code atom execution. It picks between two mutually-exclusive execution
+ * modes based on the atom's `interpreterVersion`:
+ *
+ *   0 (JavaScript / legacy): delegated to `./code-legacy.ts`. That path
+ *     exposes the classic wrapped* helpers (Move, Chamfer, CutAssembly, ...)
+ *     and the `userLib` abstraction. Do not add new features there.
+ *
+ *   1+ (TypeScript): handled in this file. The user authored a `run(...)`
+ *     function that was transpiled to JS on the main thread (see
+ *     `src/molecules/code.js#transpileTypeScript`). Our job here is to:
+ *       - prepare arguments into `AbundanceObj` / primitive form,
+ *       - invoke `run(...)` inside a sandboxed Function,
+ *       - convert the returned `AbundanceObj` (or array) back to a
+ *         RealizedAssembly for downstream cache / render consumption.
+ *     TS-mode users do NOT have access to `userLib`, wrapped* helpers, or
+ *     anything else carried over from the legacy surface — the only globals
+ *     exposed are `replicad`, `AbundanceObj`, and `AbundanceProps`.
+ */
 import * as util from "./util";
 import { AbundanceObject } from "./util";
-import * as replicad from "replicad";
 import { RequestContext } from "./geometryProvider";
+import {
+  executeCode as executeLegacy,
+  RealizedAssembly,
+  addAssemblyPartsToCache,
+  composeID,
+  ensureDimension,
+  realizeAssembly,
+  validateUserCode,
+} from "./code-legacy";
+// Raw JS source for the AbundanceObj / AbundanceProps runtime framework.
+// Prepended to every transpiled user program so that `new AbundanceObj(...)`
+// in user code resolves to the classes defined in `ts-framework.js`.
+import ABUNDANCE_TS_FRAMEWORK_JS from "./ts-framework.js?raw";
 
 /**
  * Helper function to check if a value is the NO_GEOMETRY sentinel.
- * Uses object shape detection since we can't import the actual NO_GEOMETRY object
- * (it would create a circular dependency).
  */
 function isNoGeometry(value: any): boolean {
   return value && typeof value === "object" && value.__NO_GEOMETRY__ === true;
 }
 
 /**
- * For backward compatibility reasons we allow users to call functions with
- * any of the UserGeometryObj types.
- */
-type UserGeometryObj = string | RealizedAssembly | ReplicadObject;
-
-/**
- * Provide a non-cached version of assemblies for user code execution.
- * This ensures that user code doesn't have to deal with cacheIDs
- * and can easily use replicad functions directly.
- */
-type RealizedAssembly = RealizedLeaf | RealizedBranch;
-
-type RealizedBranch = {
-  geometry: RealizedAssembly[];
-  plane: replicad.Plane;
-  dimension: "2D" | "3D" | "Wire";
-  color: string;
-  tags: string[];
-  bom: string[];
-};
-
-type RealizedLeaf = {
-  geometry: ReplicadObject[]; // For backwards compatibility this is an array. But it always contains a single item.
-  plane: replicad.Plane;
-  dimension: "2D" | "3D" | "Wire";
-  color: string;
-  tags: string[];
-  bom: string[];
-};
-
-/**
- * A best-effort check for exploits in user provided code. Checks for a series of known bad patterns.
- * @param {string} code - The JavaScript code string to validate
- * @returns {boolean} True if code appears safe, throws error if dangerous patterns detected
- */
-function validateUserCode(code: string): boolean {
-  const dangerousPatterns = [
-    /eval\s*\(/,
-    /import\s*\(/,
-    /require\s*\(/,
-    /process\s*\./,
-    /global\s*\./,
-    /window\s*\./,
-    /document\s*\./,
-    /XMLHttpRequest/,
-    /fetch\s*\(/,
-    /localStorage/,
-    /sessionStorage/,
-    /IndexedDB/,
-    /WebSocket/,
-    /Worker\s*\(/,
-    /setTimeout\s*\(/,
-    /setInterval\s*\(/,
-    /__proto__/,
-  ];
-
-  for (const pattern of dangerousPatterns) {
-    if (pattern.test(code)) {
-      throw new Error(
-        `Code contains potentially dangerous pattern: ${pattern.source}`,
-      );
-    }
-  }
-
-  return true;
-}
-
-/**
- * A function which converts any input into Abundance style geometry. Input can be a library ID, an abundance object, or a single geometry object.
- * This is useful for allowing our functions to work within the Code atom or within the flow canvas.
- */
-async function toGeometry(
-  input: UserGeometryObj,
-  name = "geometry",
-  userLib: { [key: string]: RealizedAssembly },
-  context: RequestContext,
-): Promise<AbundanceObject> {
-  if (input === undefined || input === null) {
-    throw new Error(name + " value is undefined or null");
-  }
-
-  if (typeof input === "string" || typeof input === "number") {
-    if (!userLib[input]) {
-      throw new Error(`Library ID ${input} does not exist.`);
-    }
-    return addAssemblyPartsToCache(userLib[input], context, "code-transient");
-  } else if ("geometry" in input) {
-    return addAssemblyPartsToCache(
-      input as RealizedAssembly,
-      context,
-      "code-transient",
-    );
-  }
-
-  // @ts-ignore - using _wrapped to detect if this is a raw replicad object
-  const raw_type = input?._wrapped?.$$?.ptrType?.name;
-  if (raw_type && raw_type instanceof String && raw_type.startsWith("TopoDS")) {
-    // If it's a raw geometry object, we wrap it in an abundance object
-    return {
-      dimension: "3D",
-      plane: util.XYPlane,
-      color: util.defaultColor,
-      tags: [],
-      bom: [],
-      geometry: await util.geometryProvider!.addSingularToCache(
-        input as ReplicadObject,
-        context,
-        "code-transient-",
-        [context.nextId++],
-      ),
-    };
-  } else {
-    // If it's something else, we throw an error
-    throw new Error(
-      name + " value cannot be interpreted as geometry: " + input,
-    );
-  }
-}
-
-/**
- * Check if a value is a primitive type (number, string, boolean, null, undefined).
+ * Check if a value is a primitive type.
  */
 function isPrimitive(value: any): boolean {
   return (
@@ -149,256 +58,149 @@ function isPrimitive(value: any): boolean {
 }
 
 /**
- * Executes the given code with the provided arguments list.
- * Can return geometry (AbundanceObject) or primitive values (number, string, boolean, null, undefined).
+ * Structural test for an AbundanceObj produced inside user code. We can't
+ * use `instanceof` because the class lives inside the sandboxed Function
+ * scope where user code runs, so identities differ across that boundary.
  */
-async function executeCode(
+function isAbundanceObj(value: any): boolean {
+  return !!value && value.__abundance === "AbundanceObj";
+}
+
+/**
+ * Convert a TypeScript `run()` return value (AbundanceObj, array of them, or
+ * a bare replicad shape) back into the RealizedAssembly shape that downstream
+ * cache / render code expects.
+ */
+function tsResultToRealized(value: any): any {
+  if (value == null) return value;
+  if (Array.isArray(value)) {
+    return {
+      geometry: value.map(tsResultToRealized),
+      plane: value[0]?.plane,
+      color: value[0]?.color,
+      tags: value[0]?.tags ?? [],
+      bom: value[0]?.bom ?? [],
+    };
+  }
+  if (isAbundanceObj(value)) {
+    const geom = value.geometry;
+    return {
+      geometry: Array.isArray(geom) ? geom : [geom],
+      plane: value.plane,
+      color: value.color,
+      tags: value.tags,
+      bom: value.bom,
+    };
+  }
+  return value;
+}
+
+/**
+ * Execute a TypeScript-mode Code atom. `code` is the already-transpiled JS
+ * produced by Monaco on the main thread; it defines a `run(...)` function
+ * that we invoke with the user's argument values.
+ */
+async function executeTsCode(
   code: string,
   argumentsArray: { [key: string]: any },
   context: RequestContext,
-  interpreterVersion: number = 0,
-  compiledCode?: string,
 ): Promise<AbundanceObject | number | string | boolean | null | undefined> {
   try {
-    // Validate input parameters
     if (typeof code !== "string") {
       throw new Error("Code must be a string");
+    }
+    if (code.length === 0) {
+      throw new Error(
+        "TypeScript atom is missing compiled output. Open the code editor and click Save to regenerate.",
+      );
     }
     if (code.length > 50000) {
       throw new Error("Code too long (maximum 50,000 characters)");
     }
 
-    // In TypeScript mode, we execute the transpiled JS (compiledCode) rather
-    // than the original source. Fall back to `code` if compiledCode is missing
-    // (e.g. loading a project saved before compilation was implemented).
-    const isTsMode = interpreterVersion >= 1;
-    const executableCode =
-      isTsMode && typeof compiledCode === "string" && compiledCode.length > 0
-        ? compiledCode
-        : code;
-
-    // Make a copy of the necessary entries in the library where all geometries are de-cached.
-    // This library copy will be in focus for the user. Has the side effect that direct assignments
-    // to the library by the user code will not affect the main library.
-    const userLib: { [key: string]: RealizedAssembly } = {};
-    let i = 0;
+    // Convert incoming Abundance geometry arguments into raw POJOs marked
+    // with `__isRawAbundanceObj`. The prepended framework's `__promoteInput`
+    // helper will wrap these into real AbundanceObj instances inside the
+    // sandbox before invoking `run()`.
     const argsSignature: string[] = [];
     for (const [key, value] of Object.entries(argumentsArray)) {
-      if (util.isAbundanceObject(value)) {
-        const newKey = `userlib_${i++}`;
-        userLib[newKey] = await realizeAssembly(value, context);
-        argumentsArray[key] = newKey;
-        argsSignature.push(JSON.stringify(value));
+      const actualValue = isNoGeometry(value) ? null : value;
+      if (util.isAbundanceObject(actualValue)) {
+        const realized = await realizeAssembly(actualValue, context);
+        const geom =
+          "geometry" in realized && Array.isArray((realized as any).geometry)
+            ? (realized as any).geometry.length === 1
+              ? (realized as any).geometry[0]
+              : (realized as any).geometry
+            : realized;
+        argumentsArray[key] = {
+          __isRawAbundanceObj: true,
+          geometry: geom,
+          color: (realized as any).color,
+          tags: (realized as any).tags,
+          bom: (realized as any).bom,
+          plane: (realized as any).plane,
+        };
+        argsSignature.push(JSON.stringify(actualValue));
       } else {
-        // Use original value for signature
         argsSignature.push(String(value));
+        argumentsArray[key] = actualValue;
       }
     }
 
-    // check cache for existing result (only geometry results are cached)
-    const cacheId = composeID(executableCode, argsSignature);
+    // Cache lookup on the (transpiled) code + args signature.
+    const cacheId = composeID(code, argsSignature);
     const cached = await util.geometryProvider!.getAssembly(cacheId, context);
-    if (cached) {
-      return cached;
-    }
+    if (cached) return cached;
 
     const batchId = "code-atom-" + cacheId;
     const batch: RequestContext | AbundanceObject =
       await util.geometryProvider!.startBatchOperation(context, batchId);
-
-    // Full assembly cache hit. No work to do.
-    if (util.isAbundanceObject(batch)) {
-      return batch;
-    }
+    if (util.isAbundanceObject(batch)) return batch;
 
     context = batch;
     context.nextId = 0;
 
-    // Validate code for dangerous patterns
-    // TODO: we probably want to allow some of these but still need to warn about them before executing
-    // the code molecule.
-    validateUserCode(executableCode);
+    validateUserCode(code);
 
-    // Create wrapper functions that handle library ID to geometry conversion
-    const wrappedMove = async (
-      geom: UserGeometryObj,
-      x: number,
-      y: number,
-      z: number,
-    ) => {
-      return realizeAssembly(
-        await move(
-          await toGeometry(geom, "move-geometry", userLib, context),
-          x,
-          y,
-          z,
-          context,
-        ),
-        context,
-      );
-    };
-
-    const wrappedRotate = async (
-      geom: string | RealizedAssembly,
-      x: number,
-      y: number,
-      z: number,
-    ) => {
-      return realizeAssembly(
-        await rotate(
-          await toGeometry(geom, "rotate-geometry", userLib, context),
-          x,
-          y,
-          z,
-          context,
-        ),
-        context,
-      );
-    };
-
-    const wrappedScale = async (geom: UserGeometryObj, scaleFactor: number) => {
-      return realizeAssembly(
-        await scale(
-          await toGeometry(geom, "scale-geometry", userLib, context),
-          scaleFactor,
-          context,
-        ),
-        context,
-      );
-    };
-
-    const wrappedFillet = async (geom: UserGeometryObj, radius: number) => {
-      return realizeAssembly(
-        await fillet(
-          await toGeometry(geom, "fillet-geometry", userLib, context),
-          radius,
-          context,
-        ),
-        context,
-      );
-    };
-
-    const wrappedChamfer = async (geom: UserGeometryObj, size: number) => {
-      return realizeAssembly(
-        await chamfer(
-          await toGeometry(geom, "chamfer-geometry", userLib, context),
-          size,
-          context,
-        ),
-        context,
-      );
-    };
-
-    const wrappedIntersect = async (
-      input1: UserGeometryObj,
-      input2: UserGeometryObj,
-    ) => {
-      return realizeAssembly(
-        await intersect(
-          await toGeometry(input1, "intersect-geometry1", userLib, context),
-          await toGeometry(input2, "intersect-geometry2", userLib, context),
-          context,
-        ),
-        context,
-      );
-    };
-
-    const wrappedAssembly = async (inputIDs: UserGeometryObj[]) => {
-      const ids = await Promise.all(
-        inputIDs.map(
-          async (id) =>
-            await toGeometry(id, "assembly-geometry", userLib, context),
-        ),
-      );
-      const res = await assembly(ids, context);
-      return realizeAssembly(res, context);
-    };
-
-    const wrappedCutAssembly = async (
-      input1: UserGeometryObj,
-      input2Array: UserGeometryObj[],
-    ) => {
-      return realizeAssembly(
-        await cutAssembly(
-          await toGeometry(input1, "cut-geometry1", userLib, context),
-          await Promise.all(
-            input2Array.map(
-              async (id) =>
-                await toGeometry(id, "cut-geometry", userLib, context),
-            ),
-          ),
-          context,
-        ),
-        context,
-      );
-    };
-
-    const wrappedGetBounds = async (geom: UserGeometryObj) => {
-      return util.getBounds(
-        await toGeometry(geom, "bounds-geometry", userLib, context),
-        context,
-      );
-    };
-
-    let keys1 = [
-      "Rotate",
-      "Move",
-      "Scale",
-      "Assembly",
-      "Intersect",
-      "CutAssembly",
-      "AssemblyMap",
-      "AssemblyAsIterable",
-      "GetBounds",
-      "Fillet",
-      "Chamfer",
-      "library",
-      "replicad",
-    ];
-    let inputValues = [
-      wrappedRotate,
-      wrappedMove,
-      wrappedScale,
-      wrappedAssembly,
-      wrappedIntersect,
-      wrappedCutAssembly,
-      assemblyMap,
-      assemblyAsIterable,
-      wrappedGetBounds,
-      wrappedFillet,
-      wrappedChamfer,
-      userLib, // TODO(tristan): I think we should deprecate this but it'll require passing the actual geom.
-      util.replicad,
-    ];
+    // Minimal TS-mode surface: only `replicad` is injected. AbundanceObj /
+    // AbundanceProps come from the prepended framework source. User code
+    // interacts with geometry directly (e.g. `shape.geometry.chamfer(...)`).
+    const keys: string[] = ["replicad"];
+    const values: any[] = [util.replicad];
     for (const [key, value] of Object.entries(argumentsArray)) {
-      // Sanitize parameter names to prevent injection
       if (!/^[a-zA-Z_$][a-zA-Z0-9_$]*$/.test(key)) {
         throw new Error(`Invalid parameter name: ${key}`);
       }
-      keys1.push(key);
-      inputValues.push(value);
+      keys.push(key);
+      values.push(value);
     }
 
-    // Use Function constructor instead of eval - still allows code execution but safer than eval
-    // In TS mode, the transpiled code declares `function run(...)`; we invoke it with the
-    // user-supplied argument values (in the order Monaco produced them).
-    const body = isTsMode
-      ? `${executableCode}\nreturn (async () => run(${Object.keys(argumentsArray).join(", ")}))();`
-      : `return (async () => { ${executableCode} })();`;
-    const userFunction = new Function(...keys1, body);
+    // Build the sandbox body:
+    //   1. framework declares AbundanceObj/AbundanceProps/__promoteInput
+    //   2. user code declares function run(...)
+    //   3. we invoke run() with each param routed through __promoteInput so
+    //      geometry inputs become real AbundanceObj instances.
+    const paramNames = Object.keys(argumentsArray);
+    const promotedArgs = paramNames
+      .map((n) => `__promoteInput(${n})`)
+      .join(", ");
+    const body =
+      // `${ABUNDANCE_TS_FRAMEWORK_JS}\n` +
+      `${code}\n` + `return (async () => run(${promotedArgs}))();`;
+    const userFunction = new Function(...keys, body);
 
-    // Execute with timeout protection
     const timeoutPromise = new Promise((_, reject) => {
       setTimeout(() => reject(new Error("Code execution timed out")), 120000); // 2 min timeout
     });
 
-    // Get the raw result from user code
-    const rawResult = await Promise.race([
-      userFunction(...inputValues),
+    let rawResult = await Promise.race([
+      userFunction(...values),
       timeoutPromise,
     ]);
 
-    // If the result is a primitive value or an array (e.g., point), clean up batch and return directly (don't cache)
+    rawResult = tsResultToRealized(rawResult);
+
     if (isPrimitive(rawResult) || Array.isArray(rawResult)) {
       // Clean up the warm cache without caching the result
       // Primitive and array (point) results are not cached - only geometry results are cached
@@ -406,7 +208,6 @@ async function executeCode(
       return rawResult;
     }
 
-    // Otherwise, process as geometry
     const processedResult = await ensureDimension(rawResult);
     const abundanceObj = await addAssemblyPartsToCache(
       processedResult as RealizedAssembly,
@@ -421,205 +222,17 @@ async function executeCode(
   }
 }
 
-async function ensureDimension(
-  assembly: RealizedAssembly | Promise<RealizedAssembly>,
-): Promise<RealizedAssembly> {
-  return Promise.resolve(assembly).then(async (resolvedAssembly) => {
-    if (isRealizedLeaf(resolvedAssembly)) {
-      if (
-        !("dimension" in resolvedAssembly) ||
-        resolvedAssembly.dimension === undefined
-      ) {
-        resolvedAssembly.dimension =
-          "_wrapped" in resolvedAssembly.geometry[0] ? "3D" : "2D";
-      }
-    } else {
-      resolvedAssembly.geometry = await Promise.all(
-        resolvedAssembly.geometry.map((child) => ensureDimension(child)),
-      );
-      resolvedAssembly.dimension = resolvedAssembly.geometry[0].dimension;
-    }
-    return resolvedAssembly;
-  });
-}
-
 /**
- * AssemblyMap
- *
- * Maps the given callbackFn to each leaf in the specified assembly. And returns
- * a new assembly of the same structure and metadata, but with transformed leafs.
- * If the provided assembly is a single entity, returns a transformed singular entity.
- *
- * @param {*} assemblyId
- * @param {*} callbackFn - A function that takes a leaf and returns a new leaf.
- * @returns a new assembly with the same structure and metadata as assemblyId,
- * but where each leaf is the result of applying callbackFn to the
- * corresponding leaf in the input assembly.
+ * Top-level Code atom executor. Dispatches on `interpreterVersion`.
  */
-async function assemblyMap(
-  assembly: RealizedAssembly,
-  callbackFn: (
-    node: RealizedLeaf,
-    depth: number,
-  ) => RealizedAssembly | Promise<RealizedAssembly>,
-): Promise<RealizedAssembly | undefined> {
-  try {
-    // Helper function to process nodes recursively
-    async function processNode(
-      node: RealizedAssembly,
-      depth: number,
-    ): Promise<RealizedAssembly | undefined> {
-      if (isRealizedLeaf(node)) {
-        return ensureDimension(callbackFn(node, depth));
-      }
-      // This is a branch node
-      else {
-        const newGeometry = await Promise.all(
-          node.geometry.map(async (child) => {
-            return await processNode(child, depth + 1);
-          }),
-        );
-
-        // Filter out any undefined results (in case callbackFn filters some nodes)
-        const filteredGeometry = newGeometry.filter(
-          (item) => item !== undefined,
-        );
-
-        if (filteredGeometry.length === 0) {
-          return undefined;
-        }
-
-        // Return a new node with the same metadata but transformed children
-        return {
-          geometry: filteredGeometry,
-          tags: node.tags || [],
-          color: node.color,
-          plane: node.plane,
-          bom: node.bom || [],
-          dimension: filteredGeometry[0].dimension,
-        };
-      }
-    }
-
-    // Start processing from the root
-    const result = await processNode(assembly, 0);
-    return result;
-  } catch (error) {
-    logError(error as Error, "AssemblyMap");
-    throw error;
-  }
-}
-
-async function assemblyAsIterable(assembly: RealizedAssembly) {
-  const result: RealizedAssembly[] = [];
-  const helper = (assembly: RealizedAssembly) => {
-    if (isRealizedLeaf(assembly)) {
-      result.push(assembly);
-    } else {
-      assembly.geometry.forEach((child) => helper(child));
-    }
-  };
-  helper(assembly);
-  return result;
-}
-
-function logError(error: Error, context: string) {
-  console.warn("error from context: ", context);
-  if (error instanceof SyntaxError) {
-    console.error("SyntaxError encountered:", error.message);
-  } else if (error instanceof ReferenceError) {
-    console.error("ReferenceError encountered:", error.message);
-  } else {
-    console.error("An error occurred:", error.message);
-  }
-
-  // Log additional error details if available
-  if (error.stack) {
-    console.error("Stack trace:", error.stack);
-  }
-  if ("lineNumber" in error) {
-    console.error("Line number:", error.lineNumber);
-  }
-  if ("columnNumber" in error) {
-    console.error("Column number:", error.columnNumber);
-  }
-  console.error("full error:");
-  console.error(error);
-}
-
-async function realizeAssembly(
-  assembly: AbundanceObject,
+export async function executeCode(
+  code: string,
+  argumentsArray: { [key: string]: any },
   context: RequestContext,
-): Promise<RealizedAssembly> {
-  if (util.isLeaf(assembly)) {
-    return {
-      ...assembly,
-      geometry: [await util.geometryProvider!.get(assembly.geometry, context)],
-      plane: util.asReplicadPlane(assembly.plane),
-    };
-  } else {
-    const children = await Promise.all(
-      (assembly.geometry as AbundanceObject[]).map(async (child) => {
-        return await realizeAssembly(child, context);
-      }),
-    );
-    return {
-      ...assembly,
-      geometry: children,
-      plane: util.asReplicadPlane(assembly.plane),
-      dimension: children[0].dimension,
-    };
+  interpreterVersion: number = 0,
+): Promise<AbundanceObject | number | string | boolean | null | undefined> {
+  if (interpreterVersion < 1) {
+    return executeLegacy(code, argumentsArray, context);
   }
+  return executeTsCode(code, argumentsArray, context);
 }
-
-function isRealizedLeaf(node: RealizedAssembly): node is RealizedLeaf {
-  return (
-    Array.isArray(node.geometry) &&
-    node.geometry.every((item) => "geometry" in item === false)
-  );
-}
-
-function composeID(code: string, argsSignature: string[]): string {
-  return [util.hashString(code), ...argsSignature].join("-");
-}
-
-async function addAssemblyPartsToCache(
-  assembly: RealizedAssembly,
-  context: RequestContext,
-  cacheId: string,
-): Promise<AbundanceObject> {
-  const helperFunc = async (
-    assembly: RealizedAssembly,
-  ): Promise<AbundanceObject> => {
-    if (isRealizedLeaf(assembly)) {
-      return {
-        ...assembly,
-        geometry: await util.geometryProvider!.addSingularToCache(
-          assembly.geometry[0],
-          context,
-          cacheId,
-          [context.nextId++], // Cache under the code atom's id + an offset within the result structure
-        ),
-        plane: assembly.plane
-          ? util.asSimplePlane(assembly.plane)
-          : util.XYPlane,
-      };
-    } else {
-      const children = await Promise.all(
-        (assembly.geometry as RealizedAssembly[]).map(async (child) => {
-          return await helperFunc(child);
-        }),
-      );
-      return {
-        ...assembly,
-        geometry: children,
-        plane: assembly.plane
-          ? util.asSimplePlane(assembly.plane)
-          : util.XYPlane,
-      };
-    }
-  };
-  return helperFunc(assembly);
-}
-
-export { executeCode };
